requirements.txt
torch>=2.1.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
datasets>=2.14.0  # only for downloading FineWeb
tiktoken>=0.5.0   # reference tokenizer for comparison
wandb>=0.15.0     # optional: experiment tracking
